SongKickGigsFactory renamed to GigsFactory and its implemented interface has been deleted once it is not necessary in our DDD model
